+++ src/hotspot/share/runtime/interfaceSupport.inline.hpp
@@ -48,9 +48,9 @@
 class InterfaceSupport: AllStatic {
 # ifdef ASSERT
  public:
-  static long _scavenge_alot_counter;
-  static long _fullgc_alot_counter;
-  static long _fullgc_alot_invocation;
+  static unsigned int _scavenge_alot_counter;
+  static unsigned int _fullgc_alot_counter;
+  static int _fullgc_alot_invocation;
 
   // Helper methods used to implement +ScavengeALot and +FullGCALot
   static void check_gc_alot() { if (ScavengeALot || FullGCALot) gc_alot(); }
